separate seedring to seedring.py
